# GH-120754: Disable buffering in Path.read_bytes